pyproject.toml
[project]
name = "unravel"
version = "1.0.0-beta"
description = "UNRAVEL: UN-biased high-Resolution Analysis and Validation of Ensembles using Light sheet images"
readme = "README.md"
license = { text = "GNU General Public License v3.0" }
authors = [
    { name = "Daniel Ryskamp Rijsketic", email = "danrijs@stanford.edu" }, 
    { name = "Austen Brooks Casey", email = "abcasey@stanford.edu" }, 
    { name = "Boris Dov Heifets", email = "bheifets@stanford.edu" }
]
maintainers = [
    { name = "Daniel Ryskamp Rijsketic", email = "danrijs@stanford.edu" }
]
dependencies = [
    "aicspylibczi>=3.1.2,<4.0.0",
    "antspyx>=0.4.2,<1.0.0",
    "connected_components_3d>=3.12.3,<4.0.0",
    "czifile>=2019.7.2,<2020.0.0",
    "dask>=2023.4.1,<2024.0.0",
    "fslpy>=3.18.2,<4.0.0",
    "h5py>=3.8.0,<4.0.0",
    "lxml>=4.9.3,<5.0.0",
    "matplotlib>=3.7.1,<4.0.0",
    "nibabel>=5.2.1,<6.0.0",
    "numpy>=1.25.2,<2.0.0",
    "opencv_python>=4.7.0.72,<5.0.0",
    "openpyxl>=3.1.2,<4.0.0",
    "pandas>=1.5.3,<2.0.0",
    "rich>=13.7.1,<14.0.0",
    "rich_argparse>=1.4.0,<2.0.0",
    "scipy>=1.13.1,<2.0.0",
    "seaborn>=0.13.2,<1.0.0",
    "setuptools>=65.5.0,<66.0.0",
    "statsmodels>=0.14.0,<1.0.0",
    "termcolor>=2.4.0,<3.0.0",
    "tifffile>=2023.8.30,<2024.0.0",
    "zarr>=2.14.2,<3.0.0"
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["unravel"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
relabel_nii = "image_tools.atlas.relable_nii:main"
wireframe = "image_tools.atlas.wireframe:main"
fdr_range = "cluster_stats.fdr_range:main"
fdr = "cluster_stats.fdr:main"
recursive_mirror_index = "cluster_stats.recursively_mirror_rev_cluster_indices:main"
brain_model = "cluster_stats.brain_model:main"
find_incongruent_clusters = "cluster_stats.find_incongruent_clusters:main"
sunburst = "cluster_stats.sunburst:main"
group_data = "cluster_stats.group_bilateral_data:main"
index = "cluster_stats.index:main"
legend = "cluster_stats.legend:main"
org_data = "cluster_stats.org_data:main"
prism = "cluster_stats.prism:main"
stats = "cluster_stats.stats:main"
summary = "cluster_stats.valid_clusters_summary:main"
table = "cluster_stats.table:main"
validate_clusters = "cluster_stats.validate_clusters:main"
crop_clusters = "cluster_stats.crop_clusters:main"
effect_sizes = "cluster_stats.effect_sizes.effect_sizes:main"
sex_effect_abs = "cluster_stats.effect_sizes.effect_sizes_by_sex__absolute:main"
sex_effect_rel = "cluster_stats.effect_sizes.effect_sizes_by_sex__relative:main"
h5_to_tifs = "image_io.h5_to_tifs:main"
img_io = "image_io.img_io:main"
img_to_npy = "image_io.img_to_npy:main"
nii_hd = "image_io.nii_hd:main"
nii_info = "image_io.nii_info:main"
nii_io = "image_io.nii_io:main"
nii_to_tifs = "image_io.nii_to_tifs:main"
nii_to_zarr = "image_io.nii_to_zarr:main"
reorient_nii = "image_io.reorient_nii:main"
tif_to_tifs = "image_io.tif_to_tifs:main"
zarr_to_nii = "image_io.zarr_to_nii:main"
avg = "image_tools.avg:main"
bbox = "image_tools.bbox:main"
extend = "image_tools.extend_image:main"
max = "image_tools.max:main"
pad = "image_tools.pad_img:main"
spatial_avg = "image_tools.spatial_averaging:main"
tif_DoG = "image_tools.tif_DoG:main"
tif_rb = "image_tools.tif_rb:main"
transpose_axes = "image_tools.transpose_axes:main"
unique = "image_tools.uniq_intensities:main"
rstats_summary = "region_stats.regional_cell_densities_summary:main"
rstats = "region_stats.regional_cell_densities:main"
rstats_counts = "region_stats.regional_counts:main"
rstats_vols = "region_stats.regional_volumes:main"
rstats_IF_mean = "region_stats.regional_IF_mean_intensities:main"
rstats_IF_mean_in_seg = "region_stats.regional_IF_mean_intensities_in_segmented_voxels:main"
rstats_IF_mean_summary = "region_stats.regional_IF_mean_intensities_summary:main"
check_brain_mask = "register.check_brain_mask:main"
check_reg = "register.check_reg:main"
metadata = "register.metadata:main"
prep_reg = "register.prep_reg:main"
reg = "register.reg:main"
brain_mask = "segment.brain_mask:main"
copy_tifs = "segment.copy_tifs:main"
seg = "segment.ilastik_pixel_classification:main"
agg_files_from_dirs = "utilities.aggregate_files_from_sample_dirs:main" 
agg_files_recursively = "utilities.aggregate_files_recursively:main"
clean_tifs = "utilities.clean_tif_dirs:main"
prepend_conditions = "utilities.prepend_conditions:main"
rename_ = "utilities.rename:main"
toggle = "utilities.toggle_samples:main"
apply_mask = "voxel_stats.apply_mask:main"
hemi_to_avg = "voxel_stats.hemi_to_LR_avg:main"
whole_to_avg = "voxel_stats.whole_to_LR_avg:main"
mirror = "voxel_stats.mirror:main"
prep_vstats = "voxel_stats.prep_vstats:main"
vstats = "voxel_stats.vstats:main"
z_score = "voxel_stats.z_score:main"
to_atlas = "warp.to_atlas:main"
to_native = "warp.to_native:main"
warp = "warp.warp:main"
